+++ runscope/resource_runscope_step.go
@@ -98,29 +98,6 @@ func resourceRunscopeStep() *schema.Resource {
 				Optional: true,
 				ForceNew: false,
 			},
-			// TODO: rename to "variable" for better UX
-			"variables": {
-				Type: schema.TypeSet,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"name": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"property": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"source": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-					},
-				},
-				Optional:      true,
-				ConflictsWith: []string{"variable"},
-				Deprecated:    "use variable instead",
-			},
 			"variable": {
 				Type: schema.TypeSet,
 				Elem: &schema.Resource{
@@ -139,35 +116,7 @@ func resourceRunscopeStep() *schema.Resource {
 						},
 					},
 				},
-				Optional:      true,
-				ConflictsWith: []string{"variables"},
-			},
-			// TODO: rename to "assertion" for better UX
-			"assertions": {
-				Type: schema.TypeList,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"source": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"property": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"comparison": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"value": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-					},
-				},
-				Optional:      true,
-				ConflictsWith: []string{"assertion"},
-				Deprecated:    "use assertion instead",
+				Optional: true,
 			},
 			"assertion": {
 				Type: schema.TypeList,
@@ -191,27 +140,7 @@ func resourceRunscopeStep() *schema.Resource {
 						},
 					},
 				},
-				Optional:      true,
-				ConflictsWith: []string{"assertions"},
-			},
-			// TODO: rename to "header" for better UX
-			"headers": {
-				Type:     schema.TypeSet,
 				Optional: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"header": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"value": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-					},
-				},
-				ConflictsWith: []string{"header"},
-				Deprecated:    "use header instead",
 			},
 			"header": {
 				Type:     schema.TypeSet,
@@ -228,7 +157,6 @@ func resourceRunscopeStep() *schema.Resource {
 						},
 					},
 				},
-				ConflictsWith: []string{"headers"},
 			},
 			"auth": {
 				Type:     schema.TypeSet,
@@ -317,21 +245,9 @@ func resourceStepRead(d *schema.ResourceData, meta interface{}) error {
 	d.Set("method", step.Method)
 	d.Set("url", step.URL)
 	d.Set("body", step.Body)
-	if _, ok := d.GetOk("variables"); ok {
-		d.Set("variables", readVariables(step.Variables))
-	} else {
-		d.Set("variable", readVariables(step.Variables))
-	}
-	if _, ok := d.GetOk("assertions"); ok {
-		d.Set("assertions", readAssertions(step.Assertions))
-	} else {
-		d.Set("assertion", readAssertions(step.Assertions))
-	}
-	if _, ok := d.GetOk("headers"); ok {
-		d.Set("headers", readHeaders(step.Headers))
-	} else {
-		d.Set("header", readHeaders(step.Headers))
-	}
+	d.Set("variable", readVariables(step.Variables))
+	d.Set("assertion", readAssertions(step.Assertions))
+	d.Set("header", readHeaders(step.Headers))
 	d.Set("scripts", step.Scripts)
 	d.Set("before_scripts", step.BeforeScripts)
 	d.Set("note", step.Note)
@@ -356,12 +272,9 @@ func resourceStepUpdate(d *schema.ResourceData, meta interface{}) error {
 	}
 
 	if d.HasChange("url") ||
-		d.HasChange("variables") ||
 		d.HasChange("variable") ||
 		d.HasChange("assertion") ||
-		d.HasChange("assertions") ||
 		d.HasChange("header") ||
-		d.HasChange("headers") ||
 		d.HasChange("body") ||
 		d.HasChange("note") {
 		client := meta.(*runscope.Client)
@@ -407,16 +320,7 @@ func createStepFromResourceData(d *schema.ResourceData) (*runscope.TestStep, str
 		step.URL = attr.(string)
 	}
 
-	func() {
-		var attr interface{}
-		var ok bool
-
-		if attr, ok = d.GetOk("variables"); !ok {
-			if attr, ok = d.GetOk("variable"); !ok {
-				return
-			}
-		}
-
+	if attr, ok := d.GetOk("variable"); ok {
 		variables := []*runscope.Variable{}
 		items := attr.(*schema.Set)
 		for _, x := range items.List() {
@@ -430,7 +334,7 @@ func createStepFromResourceData(d *schema.ResourceData) (*runscope.TestStep, str
 			variables = append(variables, &variable)
 		}
 		step.Variables = variables
-	}()
+	}
 
 	if v, _ := d.GetOk("auth"); v != nil {
 		authSet := v.(*schema.Set).List()
@@ -444,16 +348,7 @@ func createStepFromResourceData(d *schema.ResourceData) (*runscope.TestStep, str
 		}
 	}
 
-	func() {
-		var attr interface{}
-		var ok bool
-
-		if attr, ok = d.GetOk("assertions"); !ok {
-			if attr, ok = d.GetOk("assertion"); !ok {
-				return
-			}
-		}
-
+	if attr, ok := d.GetOk("assertion"); ok {
 		assertions := []*runscope.Assertion{}
 		items := attr.([]interface{})
 		for _, x := range items {
@@ -467,20 +362,10 @@ func createStepFromResourceData(d *schema.ResourceData) (*runscope.TestStep, str
 
 			assertions = append(assertions, &variable)
 		}
-
 		step.Assertions = assertions
-	}()
+	}
 
-	func() {
-		var attr interface{}
-		var ok bool
-
-		if attr, ok = d.GetOk("headers"); !ok {
-			if attr, ok = d.GetOk("header"); !ok {
-				return
-			}
-		}
-
+	if attr, ok := d.GetOk("header"); ok {
 		step.Headers = make(map[string][]string)
 		items := attr.(*schema.Set)
 		for _, x := range items.List() {
@@ -488,7 +373,7 @@ func createStepFromResourceData(d *schema.ResourceData) (*runscope.TestStep, str
 			header := item["header"].(string)
 			step.Headers[header] = append(step.Headers[header], item["value"].(string))
 		}
-	}()
+	}
 
 	if attr, ok := d.GetOk("scripts"); ok {
 		step.Scripts = expandStringList(attr.([]interface{}))